Added function to print the final graph and fixed bugs
Added the function printDot() which prints the final graph in the DOT format,
to be used with graphviz.

Fixed bug: when adding an edge during the analysis, the vertex representative
is used to prevent adding edges to merged vertices.

Added some code to consolidate the points-to set of merged vertices after the
analysis is done. This solves potential problems with allPointsTo() and also
corrects the print() output.